// with-contentful/types/sourcebit.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface Metadata {
    __metadata: {
        id: string;
        modelName: string;
        urlPath?: string;
    };
}

export interface IButton extends Metadata {
    /** Label */
    label: string;

    /** URL */
    url?: string | undefined;

    /** Color scheme */
    theme: 'primary' | 'secondary';
}


export interface ICard extends Metadata {
    /** Heading */
    heading: string;

    /** Subheading */
    subheading?: string | undefined;

    /** URL */
    url?: string | undefined;
}


export interface ICardGridSection extends Metadata {
    /** Heading */
    heading: string;

    /** Subheading */
    subheading?: string | undefined;

    /** Cards */
    cards: ICard[];
}


export interface IFooterConfig extends Metadata {
    /** Body */
    body: string;
}


export interface IHeroSection extends Metadata {
    /** Heading */
    heading?: string | undefined;

    /** Subheading */
    subheading?: string | undefined;

    /** Buttons */
    buttons: IButton[];
}


export interface IPage extends Metadata {
    /** Slug */
    slug: string;

    /** Layout */
    layout: string;

    /** Title */
    title: string;

    /** Sections */
    sections: (ICardGridSection | IHeroSection)[];
}


export interface ISiteConfig extends Metadata {
    /** Title */
    title: string;

    /** Favicon */
    favicon?: string | null;

    /** Footer */
    footer: IFooterConfig;
}


export type CONTENT_TYPE = 'Button' | 'Card' | 'CardGridSection' | 'FooterConfig' | 'HeroSection' | 'Page' | 'SiteConfig';

export type LOCALE_CODE = 'en-US';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US';